pyproject.toml
[tool.poetry]
name = "isocial"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.98.0"}
requests = "^2.31.0"
openai = "^0.27.8"
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.19"}
aiosqlite = "^0.19.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^11.0.0"}
google-api-python-client = "^2.90.0"
python-dotenv = "^1.0.0"
twitterapi = "^2.8.2"
pandas = "^2.0.2"
tiktokapi = "^5.2.2"
tweepy = "^4.14.0"
tiktoken = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
